requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==1.10.14
alembic==1.13.1
python-jose==3.3.0
bcrypt==4.1.2
python-multipart==0.0.9